+++ example/lib/pages/home_page.dart
@@ -38,6 +38,7 @@ class HomePage extends StatelessWidget {
     'page_setup': Icons.print_outlined,
     'grouping': Icons.account_tree_outlined,
     'comments': Icons.comment_outlined,
+    'workbook_protection': Icons.lock_person_outlined,
   };
 
   static const _showcaseIcons = <String, IconData>{